Clean up test files and remove unnecessary scripts
